requirements.txt
streamlit
python-dotenv
langchain
langchain_core
langchain_community
langchain-chroma
langchain-groq
langchain-huggingface
pypdf